*.txt: Fix spelling and grammar mistakes in documentation
This patch is the result of reviewing (by a non-foreign speaker, oh! the
irony!) the documentation provided by the upstream author in
the sources (text files).

It tries to fix some grammar and spelling mistake, while trying not
to add any new ones.

This file is distributed with the same license as the chntpwd sources